+++ src/DateTime/Clocks/SystemClock.php
@@ -1,0 +1,13 @@
+<?php
+namespace Praline\DateTime\Clocks;
+
+use Praline\DateTime\ClockInterface;
+
+// 單純地傳回系統時鐘
+class SystemClock implements ClockInterface
+{
+    public function now(): \DateTime
+    {
+        return new \DateTime();
+    }
+}